backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
email-validator==2.1.0
